fix : c언어 코드 채점 오류 수정
c언어 코드 채점 오류 수정

+++ app/schemas/task.py
@@ -27,7 +27,7 @@ class TaskMeta(BaseModel):
     id : int
     title : str
     diff : int
-    rate : int
+    rate : float
     count :int| None=None
 
 class TaskListOut(PaginationOut):
